c318: extract BenchmarksLlm tab cluster (LlmTeaLeavesSection) to own file (#441 P0 2.2)
Benchmarks.tsx 225 LOC -> 211 LOC. Benchmarks chunk 8.10 KB -> 5.14 KB (-2.96 KB).
New chunk: BenchmarksLlm-BRaF54MM.js 3.33 KB.
Removed newly-unused useQueries + Section imports + duplicate LABELLED_SCENES from Benchmarks.tsx.
ALL 6 Benchmarks tabs now extracted + lazy. typecheck clean, 44/44 tests pass.
